Bump version of framework-test.
I considered switching to pulling in *all* jar files in that directory,
with the hope that there would always be exactly one (from the most
recent build). However, I'm not sure if that's guaranteed, and I didn't
want to produce a mysterious problem under which the newest jar isn't
seen because Gradle/javac sees an older jar first....

So, for now, we'll have to bump the version every time we pull in a CF
after one of its releases. Maybe the ideal fix would be for CF to
produce a jar without a version number, at least optionally?