To look for annotations on outer types, examine Tree, not AnnotatedTypeMirror.
This solves a number of issues:

- It (I think) avoids reporting errors about types declared in
dependencies. (This issue was noted in a TODO in the code.) I haven't
tested for this, though.

- It avoids reporting errors for `foo.new Bar()` when `foo` has
unspecified nullness:
https://github.com/jspecify/jspecify/commit/d1cd6b2aedfdce68cbe231ac00d1714d4fb9a989#diff-4e11c78e6f98776560db212d945721d3700a49d678d872016edff5dc76ef0097R50
This was prompted by
https://github.com/google/turbine/blob/130009a7f10188b9f6c334727d65ff2c50055e85/java/com/google/turbine/binder/JimageClassBinder.java#L80

- It gives us access to annotations on outer types when the type appears
as a type argument (or something like that). I think this is part of
what the end of the existing comment in visitMemberSelect is getting at,
except that I think CF *itself* is (wrongly, I think) calling
AnnotatedTypeFactory.fromElement while processing such types. Anyway,
this commit fixes our handling of samples like
https://github.com/jspecify/jspecify/blob/d1cd6b2aedfdce68cbe231ac00d1714d4fb9a989/samples/AnnotatedInnerOfNonParameterized.java#L43